move LlmConfig and use defaults more
